backend/requirements.txt
alabaster==1.0.0
asgiref==3.8.1
babel==2.16.0
certifi==2024.12.14
charset-normalizer==3.4.0
colorama==0.4.6
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docutils==0.21.2
idna==3.10
imagesize==1.4.1
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pillow==10.4.0
Pygments==2.18.0
PyJWT==2.10.0
PyYAML==6.0.2
requests==2.32.3
snowballstemmer==2.2.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.3